some style changes
