+++ ui/src/index.css
@@ -11,11 +11,11 @@
 @layer components {
     .main-grid {
         display: grid;
-        grid-template-columns: auto 1fr auto auto auto;
-        @apply gap-x-5 gap-y-0.5;
+        grid-template-columns: 1fr auto auto auto;
+        @apply gap-x-5 gap-y-1;
     }
 
     .main-grid-header {
-        @apply mb-2 font-semibold
+        @apply mb-3 text-gray-400 font-semibold;
     }
 }
